+++ app/form-components/FormStageBox.tsx
